Try to better organize templates
