+++ base/src/main/scala/clazz/Applicative.scala
@@ -9,5 +9,15 @@ trait Applicative[F[_]] {
 object Applicative {
   def apply[F[_]](implicit F: Applicative[F]): Applicative[F] = F
 
+  trait Class[F[_]] extends Template[F] with Map[F] {}
+
+  trait Template[F[_]] extends Applicative[F] with Apply.Class[F] {
+    implicit final def applicative: Applicative[F] = this
+  }
+
+  trait Map[F[_]] extends Functor[F] { self: Applicative[F] =>
+    override final def map[A, B](ma: F[A])(f: (A) => B): F[B] = apply.ap(ma)(pure(f))
+  }
+
   object syntax extends ApplicativeSyntax
 }
