--- base/src/main/scala/clazz/ApplyClass.scala
@@ -1,6 +1,0 @@
-package scato
-package clazz
-
-trait ApplyClass[F[_]] extends Apply[F] with FunctorClass[F] {
-  implicit final def apply: Apply[F] = this
-}
