--- base/src/main/scala/clazz/FoldableClass.scala
@@ -1,6 +1,0 @@
-package scato
-package clazz
-
-trait FoldableClass[F[_]] extends Foldable[F]{
-  implicit final def foldable: Foldable[F] = this
-}
