--- base/src/main/scala/clazz/FunctorClass.scala
@@ -1,6 +1,0 @@
-package scato
-package clazz
-
-trait FunctorClass[F[_]] extends Functor[F]{
-  implicit final def functor: Functor[F] = this
-}
