+++ base/src/main/scala/clazz/MonadInstances.scala
@@ -2,7 +2,7 @@ package scato
 package clazz
 
 trait MonadInstances {
-  implicit val identity: Monad[Identity] = new MonadTemplate[Identity] {
+  implicit val identity: Monad[Identity] = new Monad.Template[Identity] {
     override def map[A, B](fa: Identity[A])(f: A => B): Identity[B] = Identity(f(fa.run))
     override def ap[A, B](fa: Identity[A])(f: Identity[A => B]): Identity[B] = Identity(f.run.apply(fa.run))
     override def pure[A](a: A): Identity[A] = Identity(a)
