+++ base/src/main/scala/clazz/MonadInstancesStdLib.scala
@@ -2,14 +2,14 @@ package scato
 package clazz
 
 trait MonadInstancesStdLib {
-  implicit val option: Monad[Option] = new MonadTemplate[Option] {
+  implicit val option: Monad[Option] = new Monad.Template[Option] {
     override def ap[A, B](oa: Option[A])(f: Option[A => B]): Option[B] = oa.flatMap(a => f.map(_(a)))
     override def flatMap[A, B](oa: Option[A])(f: A => Option[B]): Option[B] = oa.flatMap(f)
     override def map[A, B](oa: Option[A])(f: A => B): Option[B] = oa.map(f)
     override def pure[A](a: A): Option[A] = Option(a)
   }
 
-  implicit val list: Monad[List] = new MonadTemplate[List] {
+  implicit val list: Monad[List] = new Monad.Template[List] {
     override def ap[A, B](xs: List[A])(f: List[A => B]): List[B] = xs.flatMap(a => f.map(_(a)))
     override def flatMap[A, B](xs: List[A])(f: A => List[B]): List[B] = xs.flatMap(f)
     override def map[A, B](xs: List[A])(f: A => B): List[B] = xs.map(f)
