--- base/src/main/scala/clazz/TraversableClass.scala
@@ -1,6 +1,0 @@
-package scato
-package clazz
-
-trait TraversableClass[F[_]] extends Traversable[F] with FunctorClass[F] with FoldableClass[F] {
-  implicit final def traversable: Traversable[F] = this
-}
