+++ base/src/main/scala/data/DisjunctionInstances.scala
@@ -5,7 +5,7 @@ import clazz._
 import Disjunction.{\/, L_, R_}
 
 trait DisjunctionInstances {
-  implicit def monad[L]: Monad[L \/ ?] = new MonadTemplate[L \/ ?] {
+  implicit def monad[L]: Monad[L \/ ?] = new Monad.Template[L \/ ?] {
 
     override def map[A, B](ma: L \/ A)(f: A => B): L \/ B =
       ma.fold[L \/ B](l => L_(l))(r => R_(f(r)))
